Encapsulate field parsing
